+++ docker/main/startsystem_fcgi.sh
@@ -31,11 +31,11 @@
 #
 # ***** END LICENSE BLOCK *****
 
-# Clixon start script without restconf. NOTE TESTS are very restricted
+# Clixon startscript for fcgi/nginx and http
 # This script is copied into the container on build time and runs
 # _inside_ the container at start in runtime. It gets environment variables
 # from the start.sh script.
-# It starts a backend and the sleeps
+# It starts a backend, a restconf daemon,  a nginx daemon,  exposes ports and the sleeps
 # for restconf.
 # See also Dockerfile of the example
 # Log msg, see with docker logs
@@ -46,21 +46,49 @@ set -eux
 
 DBG=${DBG:-0}
 
+WWWUSER=${WWWUSER:-www-data}
+
 # Initiate clixon configuration (env variable)
 echo "$CONFIG" > /usr/local/etc/clixon.xml
 
 # Initiate running db (env variable)
 echo "$STORE" > /usr/local/var/example/running_db
 
+>&2 echo "Write nginx config files"
+# nginx site config file
+cat <<EOF > /etc/nginx/conf.d/default.conf
+#
+server {
+        listen 80 default_server;
+        listen localhost:80 default_server;
+        listen [::]:80 default_server;
+	server_name localhost;
+	server_name _;
+	location / {
+	    fastcgi_pass unix:/www-data/fastcgi_restconf.sock;
+	    include fastcgi_params;
+        }
+	location /restconf {
+	    fastcgi_pass unix:/www-data/fastcgi_restconf.sock;
+	    include fastcgi_params;
+        }
+	location /streams {
+	    fastcgi_pass unix:/www-data/fastcgi_restconf.sock;
+	    include fastcgi_params;
+ 	    proxy_http_version 1.1;
+	    proxy_set_header Connection "";
+        }
+}
+EOF
+
 # This is a clixon site test file. 
 # Add to skiplist:
 # - all 3rd party model testing (you need to download the repos)
 # - test_install.sh since you dont have the make environment
 # - test_order.sh XXX this is a bug need debugging
-# - NOTE all restconf tests skipped which makes these tests very constrained
 cat <<EOF > /usr/local/bin/test/site.sh
 # Add your local site specific env variables (or tests) here.
-SKIPLIST="test_api.sh test_c++.sh test_yangmodels.sh test_openconfig.sh test_install.sh test_privileges.sh test_augment.sh test_choice.sh test_identity.sh test_nacm_datanode_read.sh test_nacm_datanode.sh test_nacm_datanode_write.sh test_nacm_default.sh test_nacm_ext.sh test_nacm_module_read.sh test_nacm_module_write.sh test_nacm_protocol.sh test_nacm.sh test_perf.sh test_perf_state_only.sh test_perf_state.sh test_restconf2.sh test_restconf_err.sh test_restconf_jukebox.sh test_restconf_listkey.sh test_restconf_patch.sh test_restconf.sh test_restconf_startup.sh test_rpc.sh test_ssl_certs.sh test_stream.sh test_submodule.sh test_upgrade_auto.sh test_upgrade_interfaces.sh test_upgrade_repair.sh test_yang_namespace.sh"
+SKIPLIST="test_api.sh test_c++.sh test_yangmodels.sh test_openconfig.sh test_install.sh test_privileges.sh"
 #IETFRFC=
 EOF
 
@@ -68,9 +96,28 @@ EOF
 # sudo: setrlimit(RLIMIT_CORE): Operation not permitted
 echo "Set disable_coredump false" > /etc/sudo.conf
 
+
 chmod 775 /usr/local/bin/test/site.sh 
 
+if [ ! -d /run/nginx ]; then
+    mkdir /run/nginx
+fi
+
+# Start nginx
+#/usr/sbin/nginx -g 'daemon off;' -c /etc/nginx/nginx.conf
+/usr/sbin/nginx -c /etc/nginx/nginx.conf
+>&2 echo "nginx started"
+
+# Start clixon_restconf (tests will kill this)
+su -c "/www-data/clixon_restconf -l f/www-data/restconf.log -D $DBG" -s /bin/sh $WWWUSER &
+>&2 echo "clixon_restconf started"
+
+# Set grp write XXX do this when creating
+chmod g+w /www-data/fastcgi_restconf.sock
+
 # Start clixon backend (tests will kill this)
+# Note if tests start too quickly, a backend may only be running and get error when start here,
+# therefore test starts need to be delayed slightly
 /usr/local/sbin/clixon_backend -D $DBG -s running -l e # logs on docker logs
 >&2 echo "clixon_backend started"